添加完整简体中文支持

+++ Marlin/src/sd/SdFatStructs.h
@@ -16,14 +16,17 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
 #pragma once
 
 /**
- * sd/SdFatStructs.h
- *
+ * \file
+ * \brief FAT file structures
+ */
+
+/**
  * Arduino SdFat Library
  * Copyright (c) 2009 by William Greiman
  *
@@ -36,7 +39,7 @@
 
 /**
  * mostly from Microsoft document fatgen103.doc
- * https://www.microsoft.com/whdc/system/platform/firmware/fatgen.mspx
+ * http://www.microsoft.com/whdc/system/platform/firmware/fatgen.mspx
  */
 
 uint8_t const BOOTSIG0 = 0x55,          // Value for byte 510 of boot block or MBR
@@ -66,9 +69,9 @@ struct partitionTable {
    * Sector part of Cylinder-head-sector address of the first block in
    * the partition. Legal values are 1-63. Only used in old PC BIOS.
    */
-  uint8_t beginSector : 6;
+  unsigned beginSector : 6;
   /** High bits cylinder for first block in partition. */
-  uint8_t beginCylinderHigh : 2;
+  unsigned beginCylinderHigh : 2;
   /**
    * Combine beginCylinderLow with beginCylinderHigh. Legal values
    * are 0-1023.  Only used in old PC BIOS.
@@ -88,9 +91,9 @@ struct partitionTable {
    * Sector part of cylinder-head-sector address of the last sector in
    * the partition.  Legal values are 1-63. Only used in old PC BIOS.
    */
-  uint8_t endSector : 6;
+  unsigned endSector : 6;
   /** High bits of end cylinder */
-  uint8_t endCylinderHigh : 2;
+  unsigned endCylinderHigh : 2;
   /**
    * Combine endCylinderLow with endCylinderHigh. Legal values
    * are 0-1023.  Only used in old PC BIOS.
@@ -125,6 +128,7 @@ typedef struct masterBootRecord mbr_t;
  * \struct fat_boot
  *
  * \brief Boot sector for a FAT12/FAT16 volume.
+ *
  */
 struct fat_boot {
   /**
@@ -405,6 +409,7 @@ uint32_t const FSINFO_LEAD_SIG   = 0x41615252,  // 'AaRR' Lead signature for a F
  * \struct fat32_fsinfo
  *
  * \brief FSINFO sector for a FAT32 volume.
+ *
  */
 struct fat32_fsinfo {
   uint32_t  leadSignature;    // must be 0x52, 0x52, 0x61, 0x41 'RRaA'
@@ -571,7 +576,7 @@ uint8_t const DIR_NAME_0xE5     = 0x05,       // escape for name[0] = 0xE5
  *
  * \return true if the entry is for part of a long name else false.
  */
-static inline uint8_t DIR_IS_LONG_NAME(const dir_t *dir) {
+static inline uint8_t DIR_IS_LONG_NAME(const dir_t* dir) {
   return (dir->attributes & DIR_ATT_LONG_NAME_MASK) == DIR_ATT_LONG_NAME;
 }
 
@@ -584,7 +589,7 @@ uint8_t const DIR_ATT_FILE_TYPE_MASK = (DIR_ATT_VOLUME_ID | DIR_ATT_DIRECTORY);
  *
  * \return true if the entry is for a normal file else false.
  */
-static inline uint8_t DIR_IS_FILE(const dir_t *dir) {
+static inline uint8_t DIR_IS_FILE(const dir_t* dir) {
   return (dir->attributes & DIR_ATT_FILE_TYPE_MASK) == 0;
 }
 
@@ -594,7 +599,7 @@ static inline uint8_t DIR_IS_FILE(const dir_t *dir) {
  *
  * \return true if the entry is for a subdirectory else false.
  */
-static inline uint8_t DIR_IS_SUBDIR(const dir_t *dir) {
+static inline uint8_t DIR_IS_SUBDIR(const dir_t* dir) {
   return (dir->attributes & DIR_ATT_FILE_TYPE_MASK) == DIR_ATT_DIRECTORY;
 }
 
@@ -604,6 +609,6 @@ static inline uint8_t DIR_IS_SUBDIR(const dir_t *dir) {
  *
  * \return true if the entry is for a normal file or subdirectory else false.
  */
-static inline uint8_t DIR_IS_FILE_OR_SUBDIR(const dir_t *dir) {
+static inline uint8_t DIR_IS_FILE_OR_SUBDIR(const dir_t* dir) {
   return (dir->attributes & DIR_ATT_VOLUME_ID) == 0;
 }
